Title: 1까지 나누기, Time: 13ms, Memory: 1MB, Status: Passed - CodeTree

+++ 231202/1까지 나누기/divide-by-1.swift
@@ -4,10 +4,10 @@ var n = Int(readLine()!)!
 
 func solution(_ n: Int) -> Int {
     
-    var cur = Float(n)
+    var cur = Int(n)
     var temp = 0
     for i in 1...n {
-        cur = cur / Float(i)
+        cur = cur / i
         temp = i
         if cur <= 1 {
             break
